+++ .gitignore
@@ -2,9 +2,9 @@ node_modules
 .DS_store
 
 /build
-/public/build
 /server-build
 .env
+.cache
 
 /prisma/data.db
 /prisma/data.db-journal
