+++ app/routes/admin+/cache_.lru.$cacheKey.ts
@@ -1,8 +1,11 @@
 import { invariantResponse } from '@epic-web/invariant'
 import { json, type LoaderFunctionArgs } from '@remix-run/node'
-import { getAllInstances, getInstanceInfo } from 'litefs-js'
-import { ensureInstance } from 'litefs-js/remix.js'
 import { lruCache } from '#app/utils/cache.server.ts'
+import {
+	getAllInstances,
+	getInstanceInfo,
+	ensureInstance,
+} from '#app/utils/litefs.server.ts'
 import { requireUserWithRole } from '#app/utils/permissions.server.ts'
 
 export async function loader({ request, params }: LoaderFunctionArgs) {
