+++ app/routes/admin+/cache_.sqlite.tsx
@@ -1,7 +1,7 @@
-import { type ActionFunctionArgs, json, redirect } from '@remix-run/node'
-import { getInstanceInfo, getInternalInstanceDomain } from 'litefs-js'
+import { json, redirect, type ActionFunctionArgs } from '@remix-run/node'
 import { z } from 'zod'
 import { cache } from '#app/utils/cache.server.ts'
+import { getInstanceInfo } from '#app/utils/litefs.server'
 
 export async function action({ request }: ActionFunctionArgs) {
 	const { currentIsPrimary, primaryInstance } = await getInstanceInfo()
@@ -27,29 +27,4 @@ export async function action({ request }: ActionFunctionArgs) {
 		await cache.set(key, cacheValue)
 	}
 	return json({ success: true })
-}
-
-export async function updatePrimaryCacheValue({
-	key,
-	cacheValue,
-}: {
-	key: string
-	cacheValue: any
-}) {
-	const { currentIsPrimary, primaryInstance } = await getInstanceInfo()
-	if (currentIsPrimary) {
-		throw new Error(
-			`updatePrimaryCacheValue should not be called on the primary instance (${primaryInstance})}`,
-		)
-	}
-	const domain = getInternalInstanceDomain(primaryInstance)
-	const token = process.env.INTERNAL_COMMAND_TOKEN
-	return fetch(`${domain}/admin/cache/sqlite`, {
-		method: 'POST',
-		headers: {
-			Authorization: `Bearer ${token}`,
-			'Content-Type': 'application/json',
-		},
-		body: JSON.stringify({ key, cacheValue }),
-	})
 }
