+++ app/routes/users+/$username_+/notes.new.tsx
@@ -1,11 +1,12 @@
 import { json, type LoaderFunctionArgs } from '@remix-run/node'
 import { requireUserId } from '#app/utils/auth.server.ts'
-import { NoteEditor, action } from './__note-editor.tsx'
+import { NoteEditor } from './__note-editor.tsx'
+
+export { action } from './__note-editor.server.tsx'
 
 export async function loader({ request }: LoaderFunctionArgs) {
 	await requireUserId(request)
 	return json({})
 }
 
-export { action }
 export default NoteEditor
