+++ package.json
@@ -12,10 +12,10 @@
   "scripts": {
     "build": "run-s build:*",
     "build:icons": "tsx ./other/build-icons.ts",
-    "build:remix": "remix build --sourcemap",
+    "build:remix": "remix vite:build --sourcemapClient",
     "build:server": "tsx ./other/build-server.ts",
     "predev": "npm run build:icons --silent",
-    "dev": "remix dev -c \"node ./server/dev-server.js\" --manual",
+    "dev": "node ./server/dev-server.js",
     "prisma:studio": "prisma studio",
     "format": "prettier --write .",
     "lint": "eslint .",
@@ -57,11 +57,11 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@react-email/components": "0.0.11",
-    "@remix-run/css-bundle": "^2.5.0",
-    "@remix-run/express": "^2.5.0",
-    "@remix-run/node": "^2.5.0",
-    "@remix-run/react": "^2.5.0",
-    "@remix-run/server-runtime": "^2.5.0",
+    "@remix-run/css-bundle": "2.7.2",
+    "@remix-run/express": "2.7.2",
+    "@remix-run/node": "2.7.2",
+    "@remix-run/react": "2.7.2",
+    "@remix-run/server-runtime": "2.7.2",
     "@sentry/profiling-node": "^1.3.5",
     "@sentry/remix": "^7.93.0",
     "address": "^2.0.1",
@@ -110,10 +110,10 @@
   "devDependencies": {
     "@faker-js/faker": "^8.3.1",
     "@playwright/test": "^1.41.0",
-    "@remix-run/dev": "^2.5.0",
-    "@remix-run/eslint-config": "^2.5.0",
-    "@remix-run/serve": "^2.5.0",
-    "@remix-run/testing": "^2.5.0",
+    "@remix-run/dev": "2.7.2",
+    "@remix-run/eslint-config": "2.7.2",
+    "@remix-run/serve": "2.7.2",
+    "@remix-run/testing": "2.7.2",
     "@sly-cli/sly": "^1.8.0",
     "@testing-library/jest-dom": "^6.2.0",
     "@testing-library/react": "^14.1.2",
@@ -137,7 +137,6 @@
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/coverage-v8": "^1.2.1",
     "autoprefixer": "^10.4.17",
-    "chokidar": "^3.5.3",
     "enforce-unique": "^1.2.0",
     "esbuild": "^0.19.11",
     "eslint": "^8.56.0",
@@ -153,7 +152,7 @@
     "remix-flat-routes": "^0.6.4",
     "tsx": "^4.7.0",
     "typescript": "^5.3.3",
-    "vite": "^5.0.11",
+    "vite": "^5.1.4",
     "vitest": "^1.2.1"
   },
   "engines": {
