+++ server/dev-server.js
@@ -4,7 +4,7 @@ if (process.env.NODE_ENV === 'production') {
 	await import('./index.js')
 } else {
 	const command =
-		'tsx watch --clear-screen=false --ignore "app/**" --ignore "build/**" --ignore "node_modules/**" --inspect ./index.js'
+		'tsx watch --clear-screen=false --ignore ".cache/**" --ignore "app/**" --ignore "vite.config.ts.timestamp-*" --ignore "build/**" --ignore "node_modules/**" --inspect ./index.js'
 	execa(command, {
 		stdio: ['ignore', 'inherit', 'inherit'],
 		shell: true,
