+++ server/index.ts
@@ -1,15 +1,6 @@
 import crypto from 'crypto'
-import path from 'path'
-import { fileURLToPath } from 'url'
-import {
-	createRequestHandler as _createRequestHandler,
-	type RequestHandler,
-} from '@remix-run/express'
-import {
-	broadcastDevReady,
-	installGlobals,
-	type ServerBuild,
-} from '@remix-run/node'
+import { createRequestHandler as _createRequestHandler } from '@remix-run/express'
+import { type ServerBuild, installGlobals } from '@remix-run/node'
 import * as Sentry from '@sentry/remix'
 import { ip as ipAddress } from 'address'
 import chalk from 'chalk'
@@ -25,19 +16,19 @@ installGlobals()
 
 const MODE = process.env.NODE_ENV
 
-const createRequestHandler = Sentry.wrapExpressCreateRequestHandler(
-	_createRequestHandler,
-)
+const createRequestHandler =
+	MODE === 'production'
+		? Sentry.wrapExpressCreateRequestHandler(_createRequestHandler)
+		: _createRequestHandler
 
-const BUILD_PATH = '../build/index.js'
-const WATCH_PATH = '../build/version.txt'
-
-/**
- * Initial build
- * @type {ServerBuild}
- */
-const build = await import(BUILD_PATH)
-let devBuild = build
+const viteDevServer =
+	MODE === 'production'
+		? undefined
+		: await import('vite').then(vite =>
+				vite.createServer({
+					server: { middlewareMode: true },
+				}),
+			)
 
 const app = express()
 
@@ -79,23 +70,27 @@ app.disable('x-powered-by')
 app.use(Sentry.Handlers.requestHandler())
 app.use(Sentry.Handlers.tracingHandler())
 
-// Remix fingerprints its assets so we can cache forever.
-app.use(
-	'/build',
-	express.static('public/build', { immutable: true, maxAge: '1y' }),
-)
+if (viteDevServer) {
+	app.use(viteDevServer.middlewares)
+} else {
+	// Remix fingerprints its assets so we can cache forever.
+	app.use(
+		'/assets',
+		express.static('build/client/assets', { immutable: true, maxAge: '1y' }),
+	)
 
-// Everything else (like favicon.ico) is cached for an hour. You may want to be
-// more aggressive with this caching.
-app.use(express.static('public', { maxAge: '1h' }))
+	// Everything else (like favicon.ico) is cached for an hour. You may want to be
+	// more aggressive with this caching.
+	app.use(express.static('build/client', { maxAge: '1h' }))
+}
 
-app.get(['/build/*', '/img/*', '/fonts/*', '/favicons/*'], (req, res) => {
+app.get(['/img/*', '/favicons/*'], (req, res) => {
 	// if we made it past the express.static for these, then we're missing something.
 	// So we'll just send a 404 and won't bother calling other middleware.
 	return res.status(404).send('Not found')
 })
 
-morgan.token('url', (req, res) => decodeURIComponent(req.url ?? ''))
+morgan.token('url', req => decodeURIComponent(req.url ?? ''))
 app.use(
 	morgan('tiny', {
 		skip: (req, res) =>
@@ -199,18 +194,27 @@ app.use((req, res, next) => {
 	return generalRateLimit(req, res, next)
 })
 
-function getRequestHandler(build: ServerBuild): RequestHandler {
-	function getLoadContext(_: any, res: any) {
-		return { cspNonce: res.locals.cspNonce }
-	}
-	return createRequestHandler({ build, mode: MODE, getLoadContext })
+async function getBuild() {
+	const build = viteDevServer
+		? viteDevServer.ssrLoadModule('virtual:remix/server-build')
+		: // @ts-ignore this should exist before running the server
+			// but it may not exist just yet.
+			await import('#build/server/index.js')
+	// not sure how to make this happy 🤷‍♂️
+	return build as unknown as ServerBuild
 }
 
 app.all(
 	'*',
-	MODE === 'development'
-		? (...args) => getRequestHandler(devBuild)(...args)
-		: getRequestHandler(build),
+	createRequestHandler({
+		getLoadContext: (_: any, res: any) => ({
+			cspNonce: res.locals.cspNonce,
+			serverBuild: getBuild(),
+		}),
+		mode: MODE,
+		// @sentry/remix needs to be updated to handle the function signature
+		build: MODE === 'production' ? await getBuild() : getBuild,
+	}),
 )
 
 const desiredPort = Number(process.env.PORT || 3000)
@@ -252,10 +256,6 @@ ${lanUrl ? `${chalk.bold('On Your Network:')}  ${chalk.cyan(lanUrl)}` : ''}
 ${chalk.bold('Press Ctrl+C to stop')}
 		`.trim(),
 	)
-
-	if (MODE === 'development') {
-		broadcastDevReady(build)
-	}
 })
 
 closeWithGrace(async () => {
@@ -263,24 +263,3 @@ closeWithGrace(async () => {
 		server.close(e => (e ? reject(e) : resolve('ok')))
 	})
 })
-
-// during dev, we'll keep the build module up to date with the changes
-if (MODE === 'development') {
-	async function reloadBuild() {
-		devBuild = await import(`${BUILD_PATH}?update=${Date.now()}`)
-		broadcastDevReady(devBuild)
-	}
-
-	// We'll import chokidar here so doesn't get bundled in production.
-	const chokidar = await import('chokidar')
-
-	const dirname = path.dirname(fileURLToPath(import.meta.url))
-	const watchPath = path.join(dirname, WATCH_PATH).replace(/\\/g, '/')
-
-	const buildWatcher = chokidar
-		.watch(watchPath, { ignoreInitial: true })
-		.on('add', reloadBuild)
-		.on('change', reloadBuild)
-
-	closeWithGrace(() => buildWatcher.close())
-}
