--- types/remix.env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="@remix-run/dev" />
-/// <reference types="@remix-run/node/globals" />
